refactor: match /test/unit structure to /src
