+++ ci/dev/test-unit.sh
@@ -3,7 +3,7 @@ set -euo pipefail
 
 main() {
   cd "$(dirname "$0")/../.."
-  cd test/unit/test-plugin
+  cd test/unit/node/test-plugin
   make -s out/index.js
   # We must keep jest in a sub-directory. See ../../test/package.json for more
   # information. We must also run it from the root otherwise coverage will not
