+++ test/unit/common/emitter.test.ts
@@ -1,8 +1,8 @@
 // Note: we need to import logger from the root
 // because this is the logger used in logError in ../src/common/util
-import { logger } from "../../node_modules/@coder/logger"
+import { logger } from "@coder/logger"
 
-import { Emitter } from "../../src/common/emitter"
+import { Emitter } from "../../../src/common/emitter"
 
 describe("emitter", () => {
   let spy: jest.SpyInstance
