+++ test/unit/node/cli.test.ts
@@ -3,9 +3,9 @@ import { promises as fs } from "fs"
 import * as net from "net"
 import * as os from "os"
 import * as path from "path"
-import { Args, parse, setDefaults, shouldOpenInExistingInstance, splitOnFirstEquals } from "../../src/node/cli"
-import { tmpdir } from "../../src/node/constants"
-import { paths } from "../../src/node/util"
+import { Args, parse, setDefaults, shouldOpenInExistingInstance, splitOnFirstEquals } from "../../../src/node/cli"
+import { tmpdir } from "../../../src/node/constants"
+import { paths } from "../../../src/node/util"
 
 type Mutable<T> = {
   -readonly [P in keyof T]: T[P]
